requirements.txt
discord~=1.7.3
art~=5.5
python-dotenv~=0.19.2
requests~=2.27.1
spotipy~=2.19.0